Final tweaks to code
KG commit:
- Removed commented out code
- Changed echos to HTML
- Remove unncessary spaces

+++ page-faq.php
@@ -25,26 +25,28 @@ get_header(); ?>
 	<?php the_content(); ?>
 	<div id="faq" class="faq-page">
 		<?php // check if the repeater field has rows of data
-		if( have_rows('faqs') ) :
-			while( have_rows('faqs') ): the_row();
-					echo "<h3>";
-						the_sub_field('category');
-					echo "</h3>";
-				if( have_rows('questions_answers') ):
-					while( have_rows('questions_answers') ): the_row();
-						echo "<div class='question-box'>";
-							if( get_sub_field('question') && get_sub_field('answer') ) :
-								echo "<button class='accordion'>";
-									the_sub_field('question');
-								echo "</button>";					
-								echo "<p class = 'panel'>";
-									the_sub_field('answer');
-								echo "</p>";					
-							endif;
-						echo "</div>"; // end question-box	  
-					endwhile;
-				endif;
-			endwhile;
+		if(function_exists( 'get_field' ) ) :
+			if( have_rows('faqs') ) :
+				while( have_rows('faqs') ): the_row();
+						echo "<h3>";
+							the_sub_field('category');
+						echo "</h3>";
+					if( have_rows('questions_answers') ):
+						while( have_rows('questions_answers') ): the_row(); ?>
+							<div class='question-box'>
+								<?php if( get_sub_field('question') && get_sub_field('answer') ) : ?>
+									<button class='accordion'>
+										<?php the_sub_field('question'); ?>
+									</button>					
+									<p class = 'panel'>
+										<?php the_sub_field('answer'); ?>
+									</p>				
+								<?php endif; ?>
+							</div><!-- end question-box -->
+						<?php endwhile;
+					endif;
+				endwhile;
+			endif;
 		endif;
 	endwhile; ?>
 	</div><!-- end faq -->
